LayoutTests/canvas/philip/tests/toDataURL.jpeg.quality.notnumber-expected.txt

< [index] >
toDataURL.jpeg.quality.notnumber
toDataURL with JPEG handles non-numeric quality parameters
References: toDataURL.jpeg.nan
Actual output:
Passed
